requirements.txt
Flask==2.0.2
gunicorn==20.1.0
line-bot-sdk==1.20.0
requests==2.26.0
python-dateutil==2.8.2